consolidated utility functions and tested adaspider
