+++ code/ADASpider.py
@@ -1,6 +1,5 @@
 import numpy as np
-from gradient import gradient
-from sto_grad import sto_grad
+from AdaSpiderUtil import *
 
 """
 inputs:
@@ -10,8 +9,8 @@ inputs:
 - max_iter: number of iterations
 """
 
-def ADASpider(w_0, max_iter, tx, y):
-    grads = []
+def ADASpider(w_0, tx, y, max_iter):
+    grads = []#np.empty((max_iter,len(w_0)))
     w = [w_0]
     n = len(y)
 
@@ -21,12 +20,11 @@ def ADASpider(w_0, max_iter, tx, y):
         else:
             i_t = np.random.choice(np.arange(1,n))
             t_grad = sto_grad(y, tx, w[t], i_t) - sto_grad(y, tx, w[t-1], i_t) - grads[t-1]
-
-        grad_sum = np.sum(np.linalg.norm(grads)**2)
-        gamma = 1 / (n**(1/4) * np.sqrt(np.sqrt(n) + grad_sum))
-        w_next = w[t] - gamma * t_grad
-
-        w.append(w_next)
+        
         grads.append(t_grad)
+        gamma = 1 / (n**(1/4) * np.sqrt(np.sqrt(n) + grad_sum(grads)))
+
+        w_next = w[t] - gamma * t_grad
+        w.append(w_next)
 
     return w
