--- code/sigmoid.py
@@ -1,5 +1,0 @@
-import numpy as np
-
-# compute sigmoid function
-def sigmoid(t):
-    return 1.0 / (1 + np.exp(-t))
